refactor(c1-b2): split replay command behind thin facade
