+++ crates/assay-cli/src/cli/commands/replay/mod.rs
@@ -1,0 +1,11 @@
+mod failure;
+mod flow;
+mod fs_ops;
+mod manifest;
+mod provenance;
+mod run_args;
+
+pub use flow::run;
+
+#[cfg(test)]
+mod tests;
